requirements.txt
flask
flask-cors
sqlalchemy
passlib
pyjwt
pandas
MetaTrader5
requests
schedule
python-telegram-bot
python-doteny